requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
firebase-admin==6.6.0
frozenlist==1.5.0
gcloud==0.18.3
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
jws==0.1.3
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.3
oauth2client==4.1.3
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
redis==5.2.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
upstash-redis==1.2.0
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.0
